+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # macOS Pkg Builder
 
+## 1.0.5
+- Fix signing issue with `pkg_signing_identity` parameter
+
 ## 1.0.4
 - Add support for pkg signing
   - Configured via new `pkg_signing_identity` parameter